requirements.txt
# development tools
Werkzeug==0.6.2
Django==1.8
django-compressor==2.0
django-debug-toolbar==0.9.4
django-extensions==1.7.0
django-haystack==2.4.1
django-oscar==1.3
django-webtest==1.7.7
# Testing dependencies
mock==1.0.1
WebTest==2.0.15
factory-boy==2.3.1
httpretty==0.8.0
pytest==2.5.2
pytest-cache==1.0
pytest-cov==1.6
pytest-django==2.6.2
pytest-pep8==1.0.6
lxml==3.7.2
xmltodict==0.10.2
# documentation
Sphinx==1.2.2
sphinx_rtd_theme

# Code style and coverage
flake8==0.8
coverage==3.7.1
pyflakes==1.5.0
coveralls==0.4.2

# Others
Whoosh==2.4.1
ipython
ipdb
uwsgi
